# ank-scheduler-admin/src/main/java/ank/hao/scheduler/controller/api/TaskApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.5).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package ank.hao.scheduler.controller.api;

import ank.hao.scheduler.controller.model.TaskData;
import com.fasterxml.jackson.databind.ObjectMapper;
import io.swagger.annotations.*;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.servlet.http.HttpServletRequest;
import javax.validation.Valid;
import javax.validation.constraints.*;
import java.io.IOException;
import java.util.List;
import java.util.Optional;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2019-07-11T14:07:43.697+08:00")

@Api(value = "task", description = "the task API")
public interface TaskApi {

    Logger log = LoggerFactory.getLogger(TaskApi.class);

    default Optional<ObjectMapper> getObjectMapper() {
        return Optional.empty();
    }

    default Optional<HttpServletRequest> getRequest() {
        return Optional.empty();
    }

    default Optional<String> getAcceptHeader() {
        return getRequest().map(r -> r.getHeader("Accept"));
    }

    @ApiOperation(value = "任务列表", nickname = "taskList", notes = "", response = TaskData.class, authorizations = {
        @Authorization(value = "token")
    }, tags={ "task", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "success", response = TaskData.class) })
    @RequestMapping(value = "/task/list",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<TaskData> taskList() {
        if(getObjectMapper().isPresent() && getAcceptHeader().isPresent()) {
            if (getAcceptHeader().get().contains("application/json")) {
                try {
                    return new ResponseEntity<>(getObjectMapper().get().readValue("{  \"names\" : [ \"names\", \"names\" ],  \"data\" : [ {    \"taskDesc\" : \"taskDesc\",    \"createTime\" : \"createTime\",    \"appName\" : \"appName\",    \"taskName\" : \"taskName\",    \"updateTime\" : \"updateTime\",    \"taskSource\" : 6,    \"taskId\" : 0  }, {    \"taskDesc\" : \"taskDesc\",    \"createTime\" : \"createTime\",    \"appName\" : \"appName\",    \"taskName\" : \"taskName\",    \"updateTime\" : \"updateTime\",    \"taskSource\" : 6,    \"taskId\" : 0  } ],  \"displayNames\" : [ \"displayNames\", \"displayNames\" ]}", TaskData.class), HttpStatus.NOT_IMPLEMENTED);
                } catch (IOException e) {
                    log.error("Couldn't serialize response for content type application/json", e);
                    return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
                }
            }
        } else {
            log.warn("ObjectMapper or HttpServletRequest not configured in default TaskApi interface so no example is generated");
        }
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }


    @ApiOperation(value = "获取任务所有(过滤)参数列表", nickname = "taskParams", notes = "", response = Object.class, authorizations = {
        @Authorization(value = "token")
    }, tags={ "task", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "success", response = Object.class) })
    @RequestMapping(value = "/task/params/{appName}/{taskName}",
        produces = { "application/json" }, 
        method = RequestMethod.POST)
    default ResponseEntity<Object> taskParams(@ApiParam(value = "执行器名",required=true) @PathVariable("appName") String appName,@ApiParam(value = "任务名",required=true) @PathVariable("taskName") String taskName,@ApiParam(value = "查询参数"  )  @Valid @RequestBody String body) {
        if(getObjectMapper().isPresent() && getAcceptHeader().isPresent()) {
            if (getAcceptHeader().get().contains("application/json")) {
                try {
                    return new ResponseEntity<>(getObjectMapper().get().readValue("\"{}\"", Object.class), HttpStatus.NOT_IMPLEMENTED);
                } catch (IOException e) {
                    log.error("Couldn't serialize response for content type application/json", e);
                    return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
                }
            }
        } else {
            log.warn("ObjectMapper or HttpServletRequest not configured in default TaskApi interface so no example is generated");
        }
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }

}